Renamed internal treeview components
