+++ Source/Extensions/Blazorise.TreeView/Styles/blazorise.treeview.scss
@@ -3,7 +3,7 @@
     margin-left: 1.25rem;
 }
 
-.tree-view-hidden {
+.tree-view-collapsed {
     display: none;
 }
