+++ Source/Extensions/Blazorise.TreeView/_TreeViewNodeContent.razor
@@ -1,0 +1,7 @@
+﻿@typeparam TNode
+@inherits BaseComponent
+<div class="tree-view-title">
+    <span class="@ClassNames" @onclick="@OnClick">
+        @ChildContent
+    </span>
+</div>
